`PhBaseWorkChain`: better handler for convergence
Plain lowering down the alpha mix could not work always.
The `nmix_ph` is already an improvement, leading to the usage
of more iterations into the mixing. The rate of the alpha mix change
is also increased, and below a certain threshold a multi mix approach
is tried. Afterwards, the handler finishes the strategies.