`PhBaseWorkChain`: fix `set_qpoints` step
The `set_qpoints` step in the outline of the `PhBaseWorkChain` contained several errors
incorrectly assuming that the inputs of the `PhCalculation` are found in the
`self.ctx.inputs.ph` namespace of the context. These should actually be placed in the
`self.ctx.inputs`, which is where the `BaseRestartWorkChain` expects to find the inputs
of the process class it wraps. Here we correctly assign the inputs in the context.

`Additionally, the `set_qpoints` step would assume that the `qpoints_force_parity` input
of the PhBaseWorkChain is always present. However, this is not a required input, and
hence we take this in consideration in the `set_qpoints` logic.