`seekpath_structure_analysis`: `HubbardStructureData` compatibility
The current implementation of the `seekpath_structure_analysis` calculation function
is not compatible with the `HubbardStructureData` data class. Although the function runs
without errors, the returned primitive structure is now simply a `StructureData`
instance, and hence we lose all information related to the Hubbard parameters.

Here we adapt the `seekpath_structure_analysis` calculation function so that it checks
if the provided input `structure` is a `HubbardStructureData` instance. If so, the
resulting primitive and conventional structures are converted into
`HubbardStructureData` instances with the same _onsite_ Hubbard parameters as the input
structure.

Intersite parameters are much more challenging to preserve correctly when primitivizing
the structure, since they cannot simply be defined in terms of the kinds of the
structure. Rather, they are defined on pairs of site indices and transations, which the
`get_explicit_kpoints_path` does not consider when primitivizing the structure. Hence,
the `seekpath_structure_analysis` function currently does not support intersite
parameters and will raise if the input `structure` has any defined.

Co-authored-by: AndresOrtegaGuerrero <34098967+AndresOrtegaGuerrero@users.noreply.github.com>