`PhCalculation`: add symmetry related exit codes
Fixes #1001

Some symmetry errors that ph.x can stop with are added. They are related
to how the PHonon code handles internally symmetry, which unfortunately
is slightly different from the PW code, which is used as a necessary
previous step to run ph.x. This is generally due to some interal hard-coded
thresholds. The solution is generally to run pw.x by specifying the lattice
specific ibrav, instead of the more general ibrav=0. This cannot be handled
though, as it would require to re-run the pw.x code.

Co-authored-by: Marnik Bercx <mbercx@gmail.com>