`Hubbard`: add useful utility functions (#996)
General routines to perform and handle better Hubbard parameters are
added, especially for the inter-site parameters. This functions can be used
in a number of use-cases, with special application to their initialization and
their robust definition during their self-consistent calculations.

In particular, we:

* Add the possibility of finding first (intersites) neighbours using
nearest neighbour finders via Voronoi algorithm (pymatgen).
This is helpful when using the `hp.x` code and having a structure
containing onsite Hubbard atoms with different number of neighbours.

* Add useful routines to allow for simple Hubbard parameters initialization,
by nearest neighbour analysis (pymatgen). Standardization of the lattice
and of the atomic positions can also be performed to have all the atoms
within the unit cell. The latter might be needed to run the `pw.x` binary,
which might complain about inter-sites defined beyond the 3x3x3 supercell.