🔧 Update Python support: drop v3.8 and add v3.12
Since we now effectively only support `aiida-core` v2.5 and above - which has dropped
Python 3.8 support - we can also drop Python 3.8 support.

At the same time, we add support for Python version 3.12. The latest Python version 3.13
cannot be supported yet since there is not support for it in `aiida-core`.

Finally, the compatibility matrix in the `README.md` is updated to reflect these changes.