🧪 Allow glob patterns in `retrieve_temporary` input
The `generate_calc_job_node` fixture has a `retrieve_temporary` input that allows the
user to specify a list of files in the fixture test folder that should be copied to
the temporary retrieved folder (as defined by the `_retrieve_temporary_list` class
attribute in the corresponding calculation job). Here we extend this functionality to
allow the user to specify glob patterns in the `retrieve_temporary` input.

Second, we also adapt the code to use the `pathlib` module for handling paths. While
doing so, we also revert a change in 0874d956c94bba6f2853f8a4984121330b5c77f8 that
introduced a try-except to handle the case where the XML is not in the fixture test
folder. This is no longer necessary, because the `glob` call will only return files that
are actually present in the fixture test folder.