👌 `PdosWorkChain`: Update protocols for improved accuracy
Update the Brillouin zone sampling algorithm and adjust the `deltaE` parameter for
DOS and pDOS to 0.01. The optimized tetrahedron method is now selected as recommended
by the Quantum ESPRESSO documentation for improved efficiency and accuracy.

Additionally, tightening the energy grid helps mitigate the issue of skipping semicore
states in certain cases.

Co-authored-by: Andres Ortega-Guerrero <34098967+AndresOrtegaGuerrero@users.noreply.github.com>
Co-authored-by: Michail Minotakis <mminotakis@gmail.com>