🧪 Create two pseudo families in `sssp` fixture
Currently the `sssp` fixture only creates a mock `SSSP/1.3/PBEsol/efficiency`
pseudopotential family and returns it. In order to also be able to test more stringent
protocols, we adapt the fixture to also create the `SSSP/1.3/PBEsol/precision` family.
At the end of the function, we still return the efficiency family, since this is used
in the CLI tests.