🧪 Extend `sssp` fixture to PseudoDojo
Add a `PseudoDojoFamily` to the `sssp` fixture to allow testing with this type of
pseudopotential family, e.g. for spin-orbit calculations. Since the fixture is now not
only providing `SsspFamily` pseudo potentials, we rename it to `pseudo_family`.

The cutoffs are set to a distinct value so tests can check for this when the
`PseudoDojoFamily` is meant to be used.