✨ Protocols: Add support for spin-orbit
Add support for spin-orbit calculations to the `get_builder_from_protocol()` method
of the `PwBaseWorkChain`. Typical spin-orbit calculations use non-collinear settings,
with the addition of of setting `SYSTEM.lspinorb` to `True`. Additionally, the default
pseudo-potential family is set to `PseudoDojo/0.4/PBEsol/FR/standard/upf`.

The `get_magnetization` function in the protocol utility module is also adapted to cover
`SpinType.SPIN_ORBIT`. Here, the logic of the function should be exactly the same as
for `SpinType.COLLINEAR`.

Co-authored-by: Michail Minotakis <mminotakis@gmail.com>