👌 `get_magnetization`: Replace `pseudo_family` input by `z_valences`
The current `get_magnetization` utility function has an input `pseudo_family` that
expects an `aiida-pseudo` `PseudoPotentialFamily` to extract the number of valence
electrons in the pseudo potential. However, when a user wants to use custom pseudo
potentials e.g. through the `overrides` of the `get_builder_from_protocol()` method,
it would still be beneficial to use the `get_magnetization` function to determine the
correct Quantum ESPRESSO inputs.

Here we make the function more generic by replacing the `pseudo_family` input by
`z_valences`, a simple dictionary that maps each kind name to the corresponding number
of valence electrons in the pseudo potential.

Co-authored-by: Edan Bainglass <edan.bainglass@gmail.com>