🐛 Protocols: Fix `overrides` that do not have `pw` key
In a94fbe41bbe81bf628a6bce490ae141c41150b31 we introduced some fixes for the `pseudos`
overrides, but also a bug that makes the `PwBaseWorkChain.get_builder_from_protocol()`
method fail in case overrides are provided that do not contain the `pw` key.

Here we fix this most egregious error, and add a test to makes sure overrides that only
specify the `pseudo_family` work as expected.